refactor(web): extract the file-reference seam into useFileSeamScene
PR 2 of the SpatialEditor decomposition (3789 -> 3703 lines). Moves the
LOD gate (expandedFileIds state + hysteresis thresholds + budget),
expandFileNode, fileRefLabelMap, missingRefSet/missingFileRefs, the
contentCache and the fileSeamOptions build into
use-file-seam-scene.ts, returning { fileSeamOptions, missingFileRefs }.
EXPAND_MIN_W/H stay exported: LinkEmbedLayer's shouldOffer mirrors the
same thresholds so the two LOD gates remain one decision.

Pinned scans move with the code:
- purity-guard.test.ts now asserts the single fileSeamOptions build site
  lives in the hook and that SpatialEditor builds none of its own
  (mutation-checked: a second build in the editor fails the guard);
  the every-scene-call-carries-the-seam check is unchanged.
- editor-state-surface.test.ts globs the new hook so the moved
  expandedFileIds useState stays on the ledger.

Behavior-preserving: spatial-editor jsdom 42 files / 469 tests, and the
seam-covering browser suites (canvas-embed-lod, canvas-ref-node,
drag-ghost-embed, SpatialEditor.browser - 45 tests) pass un-weakened.
