+++ apps/web/src/components/spatial-editor/SpatialEditor.tsx
@@ -100,9 +100,7 @@ import {
   recordReconnection,
   writeClipboardFragment,
 } from '../../lib/clipboard-store.js'
-import { createSpatialContentCache } from '../../lib/content-cache.js'
 import { hapticTick } from '../../lib/haptics.js'
-import { composeReferenceSeam } from '../../lib/layout-worker-protocol.js'
 import { useKeyedSvg } from '../../lib/use-keyed-svg.js'
 import type { BoxMove } from './align.js'
 import {
@@ -195,6 +193,7 @@ import {
   ToolPalette,
 } from './ToolPalette.js'
 import { MINIMAP_MIN_ROOT_WIDTH_PX, useEditorMeasurements } from './use-editor-measurements.js'
+import { EXPAND_MIN_H, EXPAND_MIN_W, useFileSeamScene } from './use-file-seam-scene.js'
 import { useGestureCaptured } from './use-gesture-captured.js'
 import { useNativeCanvasListeners } from './use-native-canvas-listeners.js'
 import { useWorkerScene } from './use-worker-scene.js'
@@ -707,105 +706,19 @@ export const SpatialEditor = forwardRef<SpatialEditorHandle, SpatialEditorProps>
       // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [canvas, externalVersion])
 
-    /**
-     * The LOD gate (embed spec v2, user decision 2026-08-08): a file node
-     * expands into an inline miniature only while its ON-SCREEN box is
-     * large enough to be legible. Hysteresis (expand at >=200x140, collapse
-     * below 160x110 CSS px) keeps pinch-zoom from flickering at the
-     * boundary, and a budget caps simultaneous miniatures at the largest
-     * candidates (deterministic tie-break by node id). The set is state so
-     * layout re-runs only when membership actually changes — never per
-     * zoom frame.
-     */
-    const EXPAND_MIN_W = 200
-    const EXPAND_MIN_H = 140
-    const COLLAPSE_MIN_W = 160
-    const COLLAPSE_MIN_H = 110
-    const EMBED_BUDGET = 8
-    const [expandedFileIds, setExpandedFileIds] = useState<ReadonlySet<string>>(new Set())
-    useEffect(() => {
-      if (resolveReference === undefined) return
-      const zoom = viewport.zoom
-      const candidates = canvas.nodes
-        .filter((node): node is Extract<SpatialNode, { type: 'file' }> => node.type === 'file')
-        .filter((node) => {
-          const w = node.width * zoom
-          const h = node.height * zoom
-          return expandedFileIds.has(node.id)
-            ? w >= COLLAPSE_MIN_W && h >= COLLAPSE_MIN_H
-            : w >= EXPAND_MIN_W && h >= EXPAND_MIN_H
-        })
-        .sort((a, b) => b.width * b.height - a.width * a.height || a.id.localeCompare(b.id))
-        .slice(0, EMBED_BUDGET)
-      const next = new Set(candidates.map((node) => node.id))
-      const unchanged =
-        next.size === expandedFileIds.size && [...next].every((id) => expandedFileIds.has(id))
-      if (!unchanged) setExpandedFileIds(next)
-    }, [canvas, viewport.zoom, resolveReference, expandedFileIds])
-    const expandFileNode = useMemo(
-      () =>
-        resolveReference === undefined
-          ? undefined
-          : (node: Extract<SpatialNode, { type: 'file' }>) => expandedFileIds.has(node.id),
-      [resolveReference, expandedFileIds],
-    )
-
-    // Opaque file references (canvas ids minted in the browser) become readable
-    // card labels through the host-supplied options list.
-    const fileRefLabelMap = useMemo(
-      () =>
-        fileRefOptions === undefined
-          ? undefined
-          : new Map(fileRefOptions.map((option) => [option.file, option.label])),
-      [fileRefOptions],
-    )
-    // The plain-data twin of the seam's `missing` field, so the worker path
-    // (which cannot take a function) sees the same missing set the
-    // synchronous and drag paths compute through the callback.
-    const missingRefSet = useMemo(() => {
-      if (missingFileRef === undefined) return undefined
-      const refs = new Set<string>()
-      for (const node of canvas.nodes) {
-        if (node.type === 'file' && missingFileRef(node.file)) refs.add(node.file)
-      }
-      return refs.size === 0 ? undefined : refs
-    }, [canvas, missingFileRef])
-    const missingFileRefs = useMemo(
-      () => (missingRefSet === undefined ? undefined : [...missingRefSet]),
-      [missingRefSet],
-    )
-    // ONE object for every render path below (committed scene, drag ghost,
-    // drag-static backdrop, resize preview). Four hand-listed copies is how a
-    // seam ends up wired into the committed render and missing from the drag
-    // overlay, which reads as content vanishing mid-gesture.
-    //
-    // `resolveReferenceContent` rides along UNCOMPOSED so the worker gate can
-    // still tell content (which cannot be serialized) from label/missing
-    // (which already cross as data).
-    // One text-node body memo per measure+theme (the cache's validity
-    // contract); rides fileSeamOptions so every synchronous render path in
-    // this component — the committed fallback, the drag backdrop, the live
-    // resize node — shares it. It never crosses to the layout worker
-    // (LayoutRequest carries plain data only); the worker keeps its own.
-    const contentCache = useMemo(
-      () => createSpatialContentCache(),
-      // eslint-disable-next-line react-hooks/exhaustive-deps -- theme and
-      // measure invalidate cached layout; nothing else does.
-      [resolvedMeasure, theme],
-    )
-    const fileSeamOptions = useMemo(
-      () => ({
-        resolveReference: composeReferenceSeam({
-          content: resolveReference,
-          labels: fileRefLabelMap,
-          missing: missingRefSet,
-        }),
-        resolveReferenceContent: resolveReference,
-        expandFileNode,
-        contentCache,
-      }),
-      [resolveReference, fileRefLabelMap, missingRefSet, expandFileNode, contentCache],
-    )
+    // The file-reference seam — the LOD gate, label/missing resolution and
+    // the content cache — built once in useFileSeamScene and spread into
+    // every scene-building call below (committed scene, drag ghost,
+    // drag-static backdrop, resize preview).
+    const { fileSeamOptions, missingFileRefs } = useFileSeamScene({
+      canvas,
+      zoom: viewport.zoom,
+      resolveReference,
+      fileRefOptions,
+      missingFileRef,
+      resolvedMeasure,
+      theme,
+    })
     // The COMMITTED scene, laid out in a worker when it can be. The drag
     // layers below keep their own synchronous paths: a gesture already has a
     // fast route through carried-side caching, and a round trip per frame
